Addition Financial-Competition2023-2024.py
This Python script is developed for the Addition Competition 2024 by Team 8. It includes functionalities for data preprocessing, 
visualization, and model training and evaluation. The script uses various libraries such as pandas, numpy, matplotlib, seaborn, sklearn, 
and xgboost to accomplish tasks like data cleaning, exploratory data analysis, feature engineering, and applying machine learning algorithms 
to predict outcomes based on the provided dataset. The goal is to develop a robust model that can accurately predict the target variable, 
contributing to the team's success in the competition.